Move storage classes to storage module
